Rename remaining src module contents
