+++ bfabric/bfabric.py
@@ -30,8 +30,8 @@ from bfabric.bfabric_config import BfabricAuth, read_config
 from bfabric.bfabric_config import BfabricConfig
 from bfabric.engine.engine_suds import EngineSUDS
 from bfabric.engine.engine_zeep import EngineZeep
-from bfabric.src.cli_formatting import HostnameHighlighter, DEFAULT_THEME
-from bfabric.src.paginator import compute_requested_pages, BFABRIC_QUERY_LIMIT, page_iter
+from bfabric.cli_formatting import HostnameHighlighter, DEFAULT_THEME
+from bfabric.utils.paginator import compute_requested_pages, BFABRIC_QUERY_LIMIT, page_iter
 from bfabric.results.result_container import ResultContainer
